# Make OffsetLog::get() thread-safe (non-mut)