--- a/benches/bench.rs
+++ b/benches/bench.rs
@@ -21,9 +21,9 @@ fn offset_log_decode(c: &mut Criterion) {
     c.bench_function("offset_log_decode", |b| {
         b.iter(|| {
             let bytes: &[u8] = &[0, 0, 0, 8, 1, 2, 3, 4, 5, 6, 7, 8, 0, 0, 0, 8, 0, 0, 0, 20];
-            let r = read_forward::<u32, _>(0, &bytes).unwrap();
+            let r = read_next::<u32, _>(0, &bytes).unwrap();
 
-            assert_eq!(&r.entry.data_buffer, &[1, 2, 3, 4, 5, 6, 7, 8]);
+            assert_eq!(&r.entry.data, &[1, 2, 3, 4, 5, 6, 7, 8]);
         })
     });
 }
@@ -110,7 +110,7 @@ fn offset_log_iter(c: &mut Criterion) {
             let log_iter = offset_log.iter();
 
             let sum: u64 = log_iter
-                .map(|val| val.data_buffer)
+                .map(|val| val.data)
                 .map(|val| from_slice(&val).unwrap())
                 .map(|val: Value| match val["value"] {
                     Value::Number(ref num) => {