--- a/src/flume_log.rs
+++ b/src/flume_log.rs
@@ -19,6 +19,6 @@ pub type Sequence = u64;
 pub trait FlumeLog {
     fn get(&self, seq: Sequence) -> Result<Vec<u8>, Error>;
     fn clear(&mut self, seq: Sequence);
-    fn latest(&self) -> Sequence;
+    fn latest(&self) -> Option<Sequence>;
     fn append(&mut self, buff: &[u8]) -> Result<Sequence, Error>;
 }