--- a/src/lib.rs
+++ b/src/lib.rs
@@ -6,12 +6,10 @@ extern crate byteorder;
 extern crate bytes;
 #[macro_use]
 extern crate failure;
-extern crate serde;
-#[macro_use]
 extern crate log;
-extern crate serde_json;
-#[macro_use]
+extern crate serde;
 extern crate serde_derive;
+extern crate serde_json;
 extern crate tokio_codec;
 
 pub mod flume_log;