I've been messing around with making some node bindings to the rusty offset log (mostly for fun/learning), and in the process I've done some refactoring. In particular, I wanted to be able to read from the OffsetLog without needing a mutable reference, and wanted to share a single File object between the OffsetLog and one or more OffsetLogIters (in a thread-safe manner). To do this, I switched the read/write calls to read_at/write_at (aka pread/pwrite) (or the windows equivalent), and removed the last_valid_offset tracking from the get() function. (Which ultimately led to more refactoring than I had originally planned).

With the reduction in system calls (seek+read vs pread), and removal of some unnecessary heap allocations, it's a bit faster now:
offset log get bench: 45ms -> 25ms
offset log append bench: 47ms -> 36ms
iterating my 600MB log: 140ms -> 100ms

Overall, the interface is mostly unchanged, but you'll have to do the following in flumeview-sql:

`OffsetLogIter::<u32, std::fs::File>::new(file)` -> `OffsetLogIter::<u32>::new(&file)`

Happy to discuss and make changes. I'll probably do more cleanup at some point, but I think it's in pretty decent shape for now.