@sbillig 2019-02-13
I pushed up a couple more commits:
1) added read_backward(offset, buf) to read the log entry prior to the specified offset (for iterating backward through the log). Also renamed read_entry to read_forward.
2) made OffsetLogIter own its File again (via File.try_clone(), which I hadn't noticed before). The &File was causing me lifetime woes.