@sbillig 2019-02-14
@pietgeursen No worries, I'm glad you approve. I'm still getting the hang of rust, so there are probably things that could be nicer. I failed to find a better solution to the read_forward/read_forward_mut split, but after some intense duckduckgoing of phrases like "abstracting over mutability", I don't think there's a better option at the moment. Also read_backward is a pretty close duplicate of read_forward, so maybe that could be cleaned up a bit.

My WIP node bindings to the offset log are here: https://github.com/sbillig/rusty-flume, btw. Simple and synchronous, and not complete yet. (I've been meaning to ask you about neon vs your node_napi library, but I'll send you a message on ssb). My short term goal is to make that work as a drop-in replacement for flumelog-offset; I sort of doubt it'll be much use to anyone, but I figured it was an easy place to get started. To make that work, I'll probably add a bidirectional (or at least backward) iterator to flumedb-rs, and maybe some other functionality. From there I'll probably try to move more node flume stuff into rustland and see if we can't manage to outweigh the cost of crossing the node/rust boundary. Or maybe I'll get sick of node and focus on the pure rust sbot or something.