@sbillig 2019-02-14
@pietgeursen A couple more changes: got rid of some of the copy pasta and made latest() useful.